[ui] split __main__.py code into separate modules
* add meshroom.ui.app.MeshroomApp to setup Meshroom UI application
* move PaletteManager to dedicated meshroom.ui.palette module
